requirements.txt
# ---------------------------------------------------------
# API Frameworks
# ---------------------------------------------------------
fastapi
uvicorn[standard]
python-multipart
pydantic
requests

# ---------------------------------------------------------
# PDF + OCR
# ---------------------------------------------------------
pymupdf==1.23.8
pillow
pytesseract
pdf2image==1.16.3
xhtml2pdf==0.2.15
reportlab==4.0.7
WeasyPrint==60.1
openpyxl==3.1.2
qrcode==7.4.2

# ---------------------------------------------------------
# AI / Embeddings / RAG
# ---------------------------------------------------------
chromadb
langchain
sentence-transformers
spacy
tqdm

# ---------------------------------------------------------
# Flask Stack
# ---------------------------------------------------------
Flask==3.0.0
Flask-Session==0.5.0
Flask-CORS==4.0.0
Bootstrap-Flask==2.2.0
Flask-SQLAlchemy==3.0.5

# ---------------------------------------------------------
# Data / Utils
# ---------------------------------------------------------
numpy==1.26.4
pandas==2.1.4

xlsxwriter